requirements.txt
asgiref==3.7.2
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
dash==2.16.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
defusedxml==0.7.1
Django==4.1.13
django-allauth==0.61.1
djongo==1.3.6
dnspython==2.6.1
Flask==3.0.2
idna==3.6
importlib_metadata==7.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
nest-asyncio==1.6.0
oauthlib==3.2.2
packaging==24.0
plotly==5.20.0
pycparser==2.21
PyJWT==2.8.0
pymongo==3.12.3
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.4.0
retrying==1.3.4
setuptools==69.2.0
six==1.16.0
sqlparse==0.2.4
tenacity==8.2.3
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
zipp==3.18.0
